Reviewer comments: rename homebrew_common to homebrew_provider
