+++ lib/puppet/provider/package/homebrew_provider.rb
